# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Lead Verification Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /results:
    get:
      operationId: listResults
      tags: [results]
      summary: List verification results with filters
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [verified, voicemail, wrong_contact, company_mismatch, invalid_number]
        - name: campaignId
          in: query
          schema:
            type: integer
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [firstName, company, phoneNumber, status, confidenceScore, calledAt]
            default: calledAt
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        "200":
          description: Paginated results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultListResponse"
  /results/export:
    get:
      operationId: exportResults
      tags: [results]
      summary: Export results as CSV data
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: campaignId
          in: query
          schema:
            type: integer
        - name: format
          in: query
          schema:
            type: string
            enum: [csv, json]
            default: csv
      responses:
        "200":
          description: Export data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportResponse"
  /results/{id}:
    get:
      operationId: getResult
      tags: [results]
      summary: Get a single result with full call detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Full result detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /call-logs:
    get:
      operationId: listCallLogs
      tags: [results]
      summary: List all call logs
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: leadId
          in: query
          schema:
            type: integer
        - name: campaignId
          in: query
          schema:
            type: integer
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [firstName, phoneNumber, humanDetected, status, callDuration, endedReason, campaignName, calledAt]
            default: calledAt
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        "200":
          description: Paginated call logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallLogListResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    ResultDetail:
      type: object
      required: [id, leadId, firstName, lastName, company, phoneNumber, status, calledAt]
      properties:
        id:
          type: integer
        leadId:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        company:
          type: string
        phoneNumber:
          type: string
        status:
          type: string
          enum: [verified, voicemail, wrong_contact, company_mismatch, invalid_number]
        callDuration:
          type: ["integer", "null"]
        aiClassification:
          type: ["string", "null"]
        transcript:
          type: ["string", "null"]
        recordingUrl:
          type: ["string", "null"]
        confidenceScore:
          type: ["number", "null"]
        calledAt:
          type: string
        campaignId:
          type: ["integer", "null"]
        campaignName:
          type: ["string", "null"]
    ResultListResponse:
      type: object
      required: [results, total, page, limit]
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/ResultDetail"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    ExportResponse:
      type: object
      required: [data, filename]
      properties:
        data:
          type: string
        filename:
          type: string
        count:
          type: integer
...
    CallLog:
      type: object
      required: [id, leadId, firstName, lastName, company, phoneNumber, status, calledAt]
      properties:
        id:
          type: integer
        leadId:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        company:
          type: string
        phoneNumber:
          type: string
        status:
          type: string
        humanDetected:
          type: ["boolean", "null"]
        endedReason:
          type: ["string", "null"]
        callSummary:
          type: ["string", "null"]
        callDuration:
          type: ["integer", "null"]
        callCost:
          type: ["number", "null"]
        confidenceScore:
          type: ["number", "null"]
        calledAt:
          type: string
        campaignName:
          type: ["string", "null"]
    CallLogListResponse:
      type: object
      required: [logs, total, page, limit]
      properties:
        logs:
          type: array
          items:
            $ref: "#/components/schemas/CallLog"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...